+++ src/_lib.styl
@@ -2,7 +2,7 @@ main-font = 'open_sanslight', sans-serif;
 sidebar-bg = #222
 sidebar-width = 300
 
-img-base = "../img"
+img-base = "../assets"
 body-bg = #4A5E71;
 body-bg-dark = rgba(#000, .4);
 menu-width = 50px;
