+++ src/app.js
@@ -1,9 +1,13 @@
-// import appcache from 'appcache';
-// appcache();
-
 import calendar from 'calendar';
 calendar.init();
 
 
 import menu from 'menu';
 menu.init();
+
+
+import router from 'router';
+router({
+	default: 'expenses',
+	handler: menu.goto
+});
