+++ src/app.styl
@@ -11,7 +11,18 @@ body
 	size 100%; overflow hidden; position fixed; // <- chrome scroll bug
 
 body
-	margin 0; font 400 15px/1 main-font; cursor default; color #fff; bg #282828 'noise.png' 0 0 repeat;
+	margin 0; font 400 15px/1 main-font; cursor default; color #fff;
+	bg #282828 'noise.png' 0 0 repeat;
 
 *
 	box-sizing border-box;
+
+
+#top
+	pos 0 0 auto 0; height top-height; padding 10;
+	// bg #182C38;
+	h1
+		pos -15 auto 0 -3; margin 0; padding 0; color rgba(#888, 0.1);
+		text-transform uppercase;
+		font-weight 900; font 75/1;
+
