--- src/data/index.js
@@ -1,8 +1,0 @@
-import categories from './categories';
-import transactions from './transactions';
-
-
-export default {
-	categories,
-	transactions
-};
