--- src/top/index.js
@@ -1,12 +1,0 @@
-// import $ from 'util';
-
-// let el;
-
-
-function init () {
-	// el = $('#top');
-}
-
-export default {
-	init
-};
