--- src/top/style.styl
@@ -1,13 +1,0 @@
-@import '_lib'
-
-
-#top
-	pos 0 0 auto 0; height top-height; padding 10;
-	// bg #182C38;
-	h1
-		pos -15 auto 0 -3; margin 0; padding 0; color rgba(#888, 0.1);
-		text-transform uppercase;
-		font-weight 900; font 75/1;
-
-	.logout
-		float right; color #fff;
